requirements.txt
watchdog
boto3
python-dotenv
configparser==5.0.2